Revert "attack controller and attack"
