+++ 2020GameProject/2020GameProject.sln
@@ -1,15 +1,18 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{ee7bbd5d-77c4-faf2-1af8-6f4e1051b85e}"
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{5DBD7BEE-C477-F2FA-1AF8-6F4E1051B85E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ee7bbd5d-77c4-faf2-1af8-6f4e1051b85e}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ee7bbd5d-77c4-faf2-1af8-6f4e1051b85e}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DBD7BEE-C477-F2FA-1AF8-6F4E1051B85E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5DBD7BEE-C477-F2FA-1AF8-6F4E1051B85E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DBD7BEE-C477-F2FA-1AF8-6F4E1051B85E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5DBD7BEE-C477-F2FA-1AF8-6F4E1051B85E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
