+++ 2020GameProject/Assets/Palettes.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e0a60c68ab6d3cf44b47a547d6e03084
+guid: 39d742c44798542b3b0353464558ef82
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
