+++ 2020GameProject/Assets/Scenes/MainScene.unity
@@ -188,24 +188,6 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-
-    - target: {fileID: 4914710595056042938, guid: cc644ba1334ea4b92bbe9bd83625967b,
-        type: 3}
-      propertyPath: attacks.Array.size
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4914710595056042938, guid: cc644ba1334ea4b92bbe9bd83625967b,
-        type: 3}
-      propertyPath: attacks.Array.data[0]
-      value: 
-      objectReference: {fileID: 4497383734388207754, guid: 4061fe785d7ad47e19c3c619e0ce153b,
-        type: 3}
-    - target: {fileID: 4914710595056042938, guid: cc644ba1334ea4b92bbe9bd83625967b,
-        type: 3}
-      propertyPath: character
-      value: 
-      objectReference: {fileID: 1116244116}
-
     - target: {fileID: 101521714087224475, guid: cc644ba1334ea4b92bbe9bd83625967b,
         type: 3}
       propertyPath: thisAnimator
@@ -216,7 +198,6 @@ PrefabInstance:
       propertyPath: m_JumpForce
       value: 280
       objectReference: {fileID: 0}
-
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: cc644ba1334ea4b92bbe9bd83625967b, type: 3}
 --- !u!1 &292079927
@@ -791,26 +772,12 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 89477103}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1116244116 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 101521714087224474, guid: cc644ba1334ea4b92bbe9bd83625967b,
-    type: 3}
-  m_PrefabInstance: {fileID: 89477103}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1116244110}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7b68a3933689e448d84aa5e243e16240, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-
 --- !u!95 &1116244114 stripped
 Animator:
   m_CorrespondingSourceObject: {fileID: 101521714087224469, guid: cc644ba1334ea4b92bbe9bd83625967b,
     type: 3}
   m_PrefabInstance: {fileID: 89477103}
   m_PrefabAsset: {fileID: 0}
-
 --- !u!1 &1572854082
 GameObject:
   m_ObjectHideFlags: 0
