--- 2020GameProject/Assets/Scripts/Attack/Attack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c8822e554b5fade4aa8b49a75f4a7bf0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
