--- 2020GameProject/Assets/Scripts/Attack/Bullets/Bullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ddcfcc565d3e54b26999b0cf43a7a91e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
