+++ 2020GameProject/Assets/Scripts/Bullets.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 898988cc9a8369b46829f7b7a710a768
+guid: 32e2d90a2f1754762a3b43e5734d336d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
