+++ 2020GameProject/Assets/Scripts/Bullets/Player_Bullet1.cs
@@ -4,13 +4,17 @@ using UnityEngine;
 
 public class Player_Bullet1 : Bullet
 {
+    public Animator animator;
     public float maxDistance;  // max distance of this bullet can reach
     public float bulletSpeed;  // speed of this bullet
 
 
+    private Vector3 direction;  // the shooting direction of this bullet
+
     // Start is called before the first frame update
     protected override void Start()
     {
+
 
         base.flyingSpeed = bulletSpeed;
         base.Start();  // call the start() in the base class
@@ -25,10 +29,20 @@ public class Player_Bullet1 : Bullet
     }
 
     /// <summary>
+    /// Function to set the shooting direction of this bullet
+    /// </summary>
+    /// <param name="direction"></param>
+    public void setDirection(Vector3 direction)
+    {
+        this.direction = direction;
+    }
+
+
+    /// <summary>
     /// This function will be called automatically when the bullet is colliding with any collider
     /// </summary>
     /// <param name="collision"></param>
-    protected override void OnTriggerEnter2D(Collider2D collision)
+    private void OnTriggerEnter2D(Collider2D collision)
     {
         if (collision.gameObject.tag == "Monster" || collision.gameObject.tag == "Ground")
         {
