--- 2020GameProject/Assets/Scripts/Player/AttackController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef9bc468bc025d84685a48409eebe3e6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
