--- 2020GameProject/Assets/Scripts/Skill/ShootingSkill.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c2bdde963d0a034898dce2b197ac4aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
