--- 2020GameProject/Assets/Scripts/Skill/Skill.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c82245df5b0f4e4083dee43b5a3d0b8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
