+++ Assets_test/Assets_test.sln
@@ -1,15 +1,18 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{800649cb-ec12-cb67-6887-3e80a5d14179}"
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{CB490680-12EC-67CB-6887-3E80A5D14179}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{800649cb-ec12-cb67-6887-3e80a5d14179}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{800649cb-ec12-cb67-6887-3e80a5d14179}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB490680-12EC-67CB-6887-3E80A5D14179}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB490680-12EC-67CB-6887-3E80A5D14179}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB490680-12EC-67CB-6887-3E80A5D14179}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB490680-12EC-67CB-6887-3E80A5D14179}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
