+++ Assets_test/Logs/Packages-Update.log
@@ -46,15 +46,3 @@ The following packages were added:
 The following packages were updated:
   com.unity.2d.animation from version 3.2.2 to 3.2.3
   com.unity.2d.spriteshape from version 3.0.11 to 3.0.12
-
-=== Thu Dec 24 15:33:23 2020
-
-Packages were changed.
-Update Mode: updateDependencies
-
-The following packages were updated:
-  com.unity.2d.animation from version 3.2.3 to 3.2.5
-  com.unity.2d.psdimporter from version 2.1.4 to 2.1.6
-  com.unity.2d.spriteshape from version 3.0.12 to 3.0.14
-  com.unity.ide.vscode from version 1.2.1 to 1.2.3
-  com.unity.test-framework from version 1.1.16 to 1.1.18
