+++ Assets_test/Packages/manifest.json
@@ -1,15 +1,15 @@
 {
   "dependencies": {
-    "com.unity.2d.animation": "3.2.5",
+    "com.unity.2d.animation": "3.2.3",
     "com.unity.2d.pixel-perfect": "2.0.4",
-    "com.unity.2d.psdimporter": "2.1.6",
+    "com.unity.2d.psdimporter": "2.1.4",
     "com.unity.2d.sprite": "1.0.0",
-    "com.unity.2d.spriteshape": "3.0.14",
+    "com.unity.2d.spriteshape": "3.0.12",
     "com.unity.2d.tilemap": "1.0.0",
     "com.unity.collab-proxy": "1.2.16",
     "com.unity.ide.rider": "1.1.4",
-    "com.unity.ide.vscode": "1.2.3",
-    "com.unity.test-framework": "1.1.18",
+    "com.unity.ide.vscode": "1.2.1",
+    "com.unity.test-framework": "1.1.16",
     "com.unity.textmeshpro": "2.1.1",
     "com.unity.timeline": "1.2.16",
     "com.unity.ugui": "1.0.0",
