+++ Assets_test/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.15f1
-m_EditorVersionWithRevision: 2019.4.15f1 (fbf367ac14e9)
+m_EditorVersion: 2019.4.9f1
+m_EditorVersionWithRevision: 2019.4.9f1 (50fe8a171dd9)
